requirements.txt
langgraph==1.0.8
pandas==3.0.1
numpy==2.4.2
yfinance==1.2.0
ta==0.11.0
requests==2.32.5
python-dotenv==1.2.1
rich==14.3.2
